requirements.txt
flask
face-recognition
numpy
Pillow
opencv-python-headless